Cleaned up many parts of the lexer interfaces
